refactor: :truck: aggiungi le importazioni dei componenti mancanti al index
