+++ frontend-quiz-app/src/components/atoms/Button.tsx
@@ -1,4 +1,4 @@
-import IconQuiz from "../molecules/IconQuiz";
+import { IconQuiz } from "../index";
 
 type ButtonProps = {
   title: string;
